decryptor function
